pyproject.toml
[project]
name = "er-scarecrow-upload"
description = "Service for uploading files to Google Drive using a service account"
readme = "README.md"
authors = [ { name = "Effective Range", email = "info@effective-range.com" } ]
requires-python = ">=3.9"
dynamic = ["version"]
dependencies = [
    "google-api-python-client>=2.90.0",
    "google-auth>=2.20.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.8.0",
    "fabric",
    "pytz",
    "retrying",
    "python-context-logger @ git+https://github.com/EffectiveRange/python-context-logger.git"
]

[build-system]
requires = [
    "setuptools>=61.0",
    "wheel",
    "setuptools_scm"
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
# Use setuptools_scm for versioning via SCM tags
version_scheme = "post-release"
local_scheme = "node-and-timestamp"
fallback_version = "0.1.0"

[project.scripts]
er-scarecrow-upload = "er_scarecrow_upload.upload:main"
er-scarecrow-fetch =  "er_scarecrow_upload.fetch:main"
er-scarecrow-fetch-upload =  "er_scarecrow_upload.fetch_upload:main"